feat: add ability to join random squad (#232)
adds a new parameter `--num-squads` and renames `--squad` to
`--squad-prefix`.

After creating or reading in the p2pool node's peer_id, it uses this to
determine which squad to join, using the pattern:
`<squad_prefix>_[0..num_squads-1]`

Also worth noting that the PeerInfo broadcast message has changed to be
a network wide topic, and seed peers will also subscribe and process
these.